"Refactored task and added pipeline configuration"

+++ chunkmydocs/src/models/server/llm.rs
@@ -1,17 +1,18 @@
 use crate::models::server::segment::SegmentType;
 use crate::utils::configs::extraction_config::Config;
+use postgres_types::{FromSql, ToSql};
 use serde::{Deserialize, Serialize};
 use utoipa::ToSchema;
 
-#[derive(Serialize, Deserialize, Debug, Clone, ToSchema)]
+#[derive(Serialize, Deserialize, Debug, Clone, ToSchema, ToSql, FromSql)]
 pub struct LLMConfig {
     pub model: LLMModel,
     pub temperature: f32,
-    pub max_tokens: usize,
+    pub max_tokens: u32,
     pub affected_segments: Vec<SegmentType>,
 }
 
-#[derive(Serialize, Deserialize, Debug, Clone, ToSchema)]
+#[derive(Serialize, Deserialize, Debug, Clone, ToSchema, ToSql, FromSql)]
 pub enum LLMModel {
     GPT4o,
     GPT4oMini,
